Azure_MCP_Server/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "atlassian-python-api>=4.0.4",
    "beautifulsoup4>=4.13.4",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.28",
    "langchain-qdrant>=0.2.0",
    "lxml>=6.0.0",
    "mcp[cli]>=1.12.2",
    "mysql-connector-python>=9.4.0",
    "openai>=1.97.1",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.0",
    "sentence-transformers>=5.0.0",
]
